Pattern_Browser/requirements-file.txt
PyQt5==5.15.6
PyYAML==6.0
